+++ src/pages/MainPage.js
@@ -1,0 +1,13 @@
+import React from "react";
+import Main from "../components/Main";
+const MainPage = () => {
+  return (
+    <div>
+      <div>header</div>
+      <Main></Main>
+      <div>footer</div>
+    </div>
+  );
+};
+
+export default MainPage;
